Improve messages class
